+++ src/snowflake/snowpark/session.py
@@ -304,6 +304,7 @@ _PYTHON_SNOWPARK_INTERNAL_TELEMETRY_ENABLED = "ENABLE_SNOWPARK_FIRST_PARTY_TELEM
 _SNOWPARK_PANDAS_DUMMY_ROW_POS_OPTIMIZATION_ENABLED = (
     "SNOWPARK_PANDAS_DUMMY_ROW_POS_OPTIMIZATION_ENABLED"
 )
+_SNOWPARK_PANDAS_HYBRID_EXECUTION_ENABLED = "SNOWPARK_PANDAS_HYBRID_EXECUTION_ENABLED"
 
 # AST encoding.
 _PYTHON_SNOWPARK_USE_AST = "PYTHON_SNOWPARK_USE_AST"
@@ -754,6 +755,14 @@ class Session:
             )
         )
 
+        from modin.config import AutoSwitchBackend
+
+        self._pandas_hybrid_execution_enabled: bool = (
+            self._conn._get_client_side_session_parameter(
+                _SNOWPARK_PANDAS_HYBRID_EXECUTION_ENABLED, AutoSwitchBackend().get()
+            )
+        )
+
         self._thread_store = create_thread_local(
             self._conn._thread_safe_session_enabled
         )
@@ -1026,6 +1035,17 @@ class Session:
         return self._dummy_row_pos_optimization_enabled
 
     @property
+    def pandas_hybrid_execution_enabled(self) -> bool:
+        """Set to ``True`` to enable hybrid execution mode (defaults to ``True``).
+        When enabled, certain operations on smaller data will automatically execute in native pandas in-memory.
+        This can significantly improve performance for operations that are more efficient in pandas than in Snowflake.
+        """
+        from modin.config import AutoSwitchBackend
+
+        self._pandas_hybrid_execution_enabled = AutoSwitchBackend().get()
+        return self._pandas_hybrid_execution_enabled
+
+    @property
     def custom_package_usage_config(self) -> Dict:
         """Get or set configuration parameters related to usage of custom Python packages in Snowflake.
 
@@ -1198,6 +1218,22 @@ class Session:
         else:
             raise ValueError(
                 "value for dummy_row_pos_optimization_enabled must be True or False!"
+            )
+
+    @pandas_hybrid_execution_enabled.setter
+    def pandas_hybrid_execution_enabled(self, value: bool) -> None:
+        """Set the value for pandas_hybrid_execution_enabled"""
+        from modin.config import AutoSwitchBackend
+
+        if value in [True, False]:
+            self._pandas_hybrid_execution_enabled = value
+            if value:
+                AutoSwitchBackend.enable()
+            else:
+                AutoSwitchBackend.disable()
+        else:
+            raise ValueError(
+                "value for pandas_hybrid_execution_enabled must be True or False!"
             )
 
     @custom_package_usage_config.setter